# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Builder API — proxies requests to local Ollama instance
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface OllamaModel {
  name: string;
  modified_at: string;
  size: number;
}

export interface ModelList {
  models: OllamaModel[];
  ollamaEndpoint: string;
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface ChatInput {
  model: string;
  messages: ChatMessage[];
  /** @nullable */
  ollamaEndpoint?: string | null;
}

export interface ChatResponse {
  message: ChatMessage;
  done: boolean;
}

export interface ProjectFile {
  path: string;
  content: string;
  language: string;
}

export interface GenerateInput {
  prompt: string;
  model: string;
  /** @nullable */
  template?: string | null;
  /** @nullable */
  ollamaEndpoint?: string | null;
}

export interface GenerateResponse {
  files: ProjectFile[];
  projectName: string;
  description: string;
}

export interface FixInput {
  files: ProjectFile[];
  errors: string;
  model: string;
  /** @nullable */
  ollamaEndpoint?: string | null;
}

export interface FixResponse {
  files: ProjectFile[];
  explanation: string;
}

export interface ImproveInput {
  files: ProjectFile[];
  model: string;
  /** @nullable */
  ollamaEndpoint?: string | null;
}

export interface ImproveResponse {
  files: ProjectFile[];
  explanation: string;
}

export interface OllamaStatus {
  connected: boolean;
  endpoint: string;
  /** @nullable */
  error?: string | null;
}