# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Builder API — proxies requests to local Ollama instance
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateInput {
  prompt: string;
  model: string;
  /** @nullable */
  template?: string | null;
  /** @nullable */
  ollamaEndpoint?: string | null;
}