# lib/api-zod/src/generated/types/generateResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Builder API — proxies requests to local Ollama instance
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectFile } from './projectFile';

export interface GenerateResponse {
  files: ProjectFile[];
  projectName: string;
  description: string;
}